Updated unix 1.
There were several small errors that needed to be corrected.
I also simplified language and improved flow.
However, the first and second half aren't as unified as I would like.

One exercise remains to be written.  It should only take a few minutes.

At the very beginnning of the lab there was some text that seemed important but was missing.
Joey should look at it and fix it.
I tried to make apparent where it was in the edited document.